requirements.txt
requests
pandas
spacy
googlemaps
rasa 
pyttsx3
openai-whispe
sounddevice 
google-cloud-texttospeech
protobuf==4.23.4
pydub
faker
pyaudio
webrtcvad
numpy==1.24

